requirements.txt
# Flask Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Google Gemini AI Integration
google-generativeai==0.3.0

# Environment and Configuration
python-dotenv==1.0.0

# CORS Support (for cross-origin requests)
flask-cors==4.0.0

# Production Server (for deployment)
gunicorn==21.2.0

# Additional Utilities
requests==2.31.0
urllib3==2.1.0

# Optional: For enhanced JSON handling
jsonschema==4.20.0

# Optional: For better error handling
sentry-sdk[flask]==1.38.0

# Development Dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0 